analysis/requirements.txt
pandas
matplotlib
seaborn
psycopg2-binary
python-dotenv